feat: version selection and command copy shortcut
